feat: divide logic to hooks and components
